tests: merge adjoints + merge FD and derivatives + clear_cache setup
