+++ tests/test_interpolation.py
@@ -1,6 +1,6 @@
 import numpy as np
 import pytest
-from conftest import skipif_yask
+from conftest import skipif_yask, unit_box, points
 from math import sin, floor
 
 from devito.cgen_utils import FLOAT
@@ -18,25 +18,6 @@ def a(shape=(11, 11)):
     yarr = np.linspace(0., 1., shape[1])
     a.data[:] = np.meshgrid(xarr, yarr)[1]
     return a
-
-
-def unit_box(name='a', shape=(11, 11)):
-    """Create a field with value 0. to 1. in each dimension"""
-    grid = Grid(shape=shape)
-    a = Function(name=name, grid=grid)
-    dims = tuple([np.linspace(0., 1., d) for d in shape])
-    a.data[:] = np.meshgrid(*dims)[1]
-    return a
-
-
-def points(grid, ranges, npoints, name='points'):
-    """Create a set of sparse points from a set of coordinate
-    ranges for each spatial dimension.
-    """
-    points = SparseFunction(name=name, grid=grid, npoint=npoints)
-    for i, r in enumerate(ranges):
-        points.coordinates.data[:, i] = np.linspace(r[0], r[1], npoints)
-    return points
 
 
 def custom_points(grid, ranges, npoints, name='points'):
@@ -248,36 +229,6 @@ def test_inject_from_field(shape, coords, result, npoints=19):
     indices = [slice(4, 6, 1) for _ in coords]
     indices[0] = slice(1, -1, 1)
     assert np.allclose(a.data[indices], result, rtol=1.e-5)
-
-
-@skipif_yask
-@pytest.mark.parametrize('shape, coords', [
-    ((11, 11), [(.05, .9), (.01, .8)]),
-    ((11, 11, 11), [(.05, .9), (.01, .8), (0.07, 0.84)])
-])
-def test_adjoint_inject_interpolate(shape, coords,
-                                    npoints=19):
-
-    a = unit_box(shape=shape)
-    a.data[:] = 0.
-    c = unit_box(shape=shape, name='c')
-    c.data[:] = 27.
-    # Inject receiver
-    p = points(a.grid, ranges=coords, npoints=npoints)
-    p.data[:] = 1.2
-    expr = p.inject(field=a, expr=p)
-    # Read receiver
-    p2 = points(a.grid, name='points2', ranges=coords, npoints=npoints)
-    expr2 = p2.interpolate(expr=c)
-    Operator(expr + expr2)(a=a, c=c)
-    # < P x, y > - < x, P^T y>
-    # Px => p2
-    # y => p
-    # x => c
-    # P^T y => a
-    term1 = np.dot(p2.data.reshape(-1), p.data.reshape(-1))
-    term2 = np.dot(c.data.reshape(-1), a.data.reshape(-1))
-    assert np.isclose((term1-term2) / term1, 0., atol=1.e-6)
 
 
 @skipif_yask
